Fixes an issue with notifyOwner system
The issue was that if for some reason, the permission at command.limits weren't exactly "administrator", then the notification wouldn't be sent. So now it supports both objects (For when using multiple permissions even though this system only works with Administrator) and toAnyCase strings.

And yes, I know that I used 2 ".toLowerCase" when I could've just made it .toUpperCase or make the "ADMINISTRATOR" string lowercase, but. I don't care :D, it's for memes